Adding Separation Lines
I'm adding in lines for separation because I think that it organizes the code more. It makes it easier for me to read the code and tell which code belongs to which area. It's about 4 lines per separation, with two lines separating the former code, 1 line noting what goes on in the next block of code, and another line to separate that title.