Major titles and Legend change
Changed major titles to be verticle, because adding in more bars will make bar names unreadable. For that, I've moved the legend to the top-left of the screen to ensure that it can be read. I understand that rotating the titles make them overlap with the bars. I'm considering lowering them, but it seems unlikely for now.